Finished reimplementation of headers.
Headers now have definite and well defined parse types. Each header is
parsed according to its key type. Tests changed to reflect new header info.
